Winter Break Updates
Few notes- didn't get as much as I wanted done, the turnPID was off when I started so I adjusted (might be different at school), wasn't sure how the stackTower method worked with the parameters so I just commented it out, and I coded it to take the first row of cubes plus the two cubes next to the tower (7 cubes which is what 2616J has). Also, the code for Game 4 was not working, only Skills 8 so that's where I coded. I assumed it was just a matter of changing the mode but we are probably going to be using the same if not similar code in both.

Karthik 1/1/2020