started 60 pointer
Got pretty far, just having trouble intaking the cube next to the tower. I  took out the blue tower and the cube in front of the tower. Saving the blue tower for the way back. ifyou have time, inake the row of 6 cubes and place cube in blue tower and stack. that was my plan.